+++ app/apis/admin/users.py
@@ -5,7 +5,8 @@ from app.util.DataBaseTools import getMaxPage, getPageContent, writeAccessLog, w
 from app.util.Response import ResponseJson
 from app.util.Request import RequestLoadJson, getClientIp
 from app.util.logger import Log
-from app.util.PasswordTools import PasswordToMd5
+from app.util.PasswordTools import PasswordToMd5, verifyPasswordRules
+from app.util.permission import groupPermission
 
 
 # 获取用户列表
@@ -36,10 +37,12 @@ def getUserList(req):
                         "lastLoginTime": item.get("lastLoginTime"),
                         "lastLoginIP": item.get("lastLoginIP"),
                         "permissionGroupID": item.get("permission_id"),
-                        "permissionGroupName": Permission_groups.objects.filter(id=item.get("permission_id")).first().name if item.get("permission_id") else None,
+                        "permissionGroupName": groupPermission(item.get("permission_id")).get_group_name() if item.get(
+                            "permission_id") else None,
                         "disable": item.get("disable")
                     })
-            writeAccessLog(req.session.get("userID"), getClientIp(req), f"Get User List(Search: {search} Page: {page} Page Size: {pageSize})")
+            writeAccessLog(req.session.get("userID"), getClientIp(req),
+                           f"Get User List(Search: {search} Page: {page} Page Size: {pageSize})")
             return ResponseJson({
                 "status": 1,
                 "data": {
@@ -50,6 +53,7 @@ def getUserList(req):
             })
     else:
         return ResponseJson({"status": -1, "msg": "请求方式不正确"})
+
 
 # 新增用户
 @Log.catch
@@ -75,8 +79,9 @@ def addUser(req):
                     return ResponseJson({"status": 0, "msg": "该用户已有账户"})
                 if email and Users.objects.filter(email=email):
                     return ResponseJson({"status": 0, "msg": "邮箱已使用过"})
-                if not re.match("^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[^\w\s]).{6,16}", password):
-                    return ResponseJson({"status": 0, "msg": "密码不符合安全要求（至少6字符，必须含有数字，小写字母，大写字母，特殊字符）"})
+                if not verifyPasswordRules(password):
+                    return ResponseJson(
+                        {"status": 0, "msg": "密码不符合安全要求（至少6字符，必须含有数字，小写字母，大写字母，特殊字符）"})
                 createUser = Users.objects.create(
                     userName=userName,
                     realName=realName,
@@ -109,19 +114,21 @@ def delUser(req):
             return ResponseJson({"status": -1, "msg": f"JSON解析失败:{e}"})
         else:
             userId = req_json.get("id")
-            Log.debug(userId)
-            Log.debug(userId == req.session.get("userID"))
             if not userId:
                 return ResponseJson({"status": -1, "msg": "参数不完整"})
             if userId == req.session.get("userID"):
-                ResponseJson({"status": 0, "msg": "不能删除当前登录用户"})
+                return ResponseJson({"status": 0, "msg": "不能删除当前登录用户"})
+            if userId == 1:
+                return ResponseJson({"status": -1, "msg": "不能删除id为1的用户"})
             query = Users.objects.filter(id=userId).first()
-            if query.permission_id and Permission_groups.objects.filter(id=query.permission_id).first().all:
-                ResponseJson({"status": 0, "msg": "不能删除拥有All权限的用户"})
+            if query.permission_id is not None and groupPermission(query.permission_id).check_group_permission("all"):
+                return ResponseJson({"status": 0, "msg": "不能删除拥有All权限的用户"})
             if query:
-                writeAudit(req.session.get("userID"), "Del User(删除用户)",
-                           "User Manager(用户管理)",
-                           f"User Id: {query.id} User Name: {query.userName}")
+                writeAudit(
+                    req.session.get("userID"),
+                    "Del User(删除用户)",
+                    "User Manager(用户管理)",
+                    f"User Id: {query.id} User Name: {query.userName}")
                 query.delete()
                 return ResponseJson({"status": 1, "msg": "用户已删除"})
             else:
@@ -145,11 +152,12 @@ def getUserPermission(req):
                 return ResponseJson({"status": -1, "msg": "参数不完整"})
             query = Users.objects.filter(id=userId).first()
             if query:
-                writeAccessLog(req.session.get("userID"), getClientIp(req),f"Get User Permission(User ID: {userId})")
+                writeAccessLog(req.session.get("userID"), getClientIp(req), f"Get User Permission(User ID: {userId})")
                 return ResponseJson({"status": 1, "data": {
                     "permissionId": query.permission_id,
-                    "permissionName": Permission_groups.objects.filter(id=query.permission_id).first().name if query.permission_id else None,
-                }})
+                    "permissionName": Permission_groups.objects.filter(
+                                        id=query.permission_id).first().name if query.permission_id else None,
+                                    }})
             else:
                 return ResponseJson({"status": 0, "msg": "用户不存在"})
     else:
@@ -178,7 +186,8 @@ def getUserInfo(req):
                     "realName": query.realName,
                     "email": query.email,
                     "permissionId": query.permission_id,
-                    "permissionName": Permission_groups.objects.filter(id=query.permission_id).first().name if query.permission_id else None,
+                    "permissionName": groupPermission(
+                        query.permission_id).get_group_name() if query.permission_id else None,
                     "disable": query.disable
                 }})
             else:
@@ -202,53 +211,75 @@ def setUserInfo(req):
             data = req_json.get("data")
             if userId and data:
                 User = Users.objects.filter(id=userId).first()
-                userName = data.get("userName")
-                realName = data.get("realName")
-                email = data.get("email")
-                password = data.get("password")
-                permission = data.get("permission")
-                disable = data.get("disable") if isinstance(data.get("disable"), bool) else bool(data.get("disable"))
+                userName: str = data.get("userName")
+                realName: str = data.get("realName")
+                email: str = data.get("email")
+                password: str = data.get("password")
+                permission: int = data.get("permission")
+                disable: bool = data.get("disable")
+                print(disable, type(disable))
                 if userName and userName != User.userName:
                     if Users.objects.filter(userName=userName):
                         return ResponseJson({"status": 0, "msg": "用户已存在"})
-                    writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Update UserName(更新用户名)",
-                               "User Manager(权限管理)",
-                               f"{User.userName}-->{userName}")
+                    writeAudit(
+                        req.session.get("userID"),
+                        "Edit User Info(编辑用户): Update UserName(更新用户名)",
+                        "User Manager(权限管理)",
+                        f"{User.userName}-->{userName}"
+                    )
                     User.userName = userName
                 if realName and realName != User.realName:
                     if Users.objects.filter(realName=realName):
                         return ResponseJson({"status": 0, "msg": "用户已存在"})
-                    writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Update RealName(更新真实姓名)",
-                               "User Manager(用户管理)",
-                               f"{User.realName}-->{realName}")
+                    writeAudit(
+                        req.session.get("userID"),
+                        "Edit User Info(编辑用户): Update RealName(更新真实姓名)",
+                        "User Manager(用户管理)",
+                        f"{User.realName}-->{realName}")
                     User.realName = realName
                 if email and email != User.email:
                     if Users.objects.filter(email=email):
                         return ResponseJson({"status": 0, "msg": "邮箱已使用"})
-                    writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Update RealName(更新邮箱)",
-                               "User Manager(用户管理)",
-                               f"{User.email}-->{email}")
+                    writeAudit(
+                        req.session.get("userID"),
+                        "Edit User Info(编辑用户): Update RealName(更新邮箱)",
+                        "User Manager(用户管理)",
+                        f"{User.email}-->{email}")
                     User.email = email
                 if password and password != User.password:
-                    if not re.match("^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[^\w\s]).{6,16}", password):
-                        return ResponseJson({"status": 0, "msg": "密码不符合安全要求（至少6字符，必须含有数字，小写字母，大写字母，特殊字符）"})
-                    writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Update RealName(更新密码)",
-                               "User Manager(用户管理)",
-                               PasswordToMd5(password))
+                    if not verifyPasswordRules(password):
+                        return ResponseJson(
+                            {
+                                "status": 0,
+                                "msg": "密码不符合安全要求（至少6字符，必须含有数字，小写字母，大写字母，特殊字符）"
+                            }
+                        )
+                    writeAudit(
+                        req.session.get("userID"),
+                        "Edit User Info(编辑用户): Update RealName(更新密码)",
+                        "User Manager(用户管理)",
+                        PasswordToMd5(password)
+                    )
                     User.password = PasswordToMd5(password)
                 if permission and permission != User.permission_id:
                     if Permission_groups.objects.filter(id=permission):
                         newPermission = Permission_groups.objects.filter(id=permission).first()
-                        writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Update Permission Group(更新权限组)",
-                                   "User Manager(权限管理)",
-                                   f"{Permission_groups.objects.filter(id=User.permission_id).first().name if User.permission_id else 'None'}-->{newPermission.name}")
+                        writeAudit(
+                            req.session.get("userID"),
+                            "Edit User Info(编辑用户): Update Permission Group(更新权限组)",
+                            "User Manager(用户管理)",
+                            f"{Permission_groups.objects.filter(id=User.permission_id).first().name if User.permission_id else 'None'}-->{newPermission.name}"
+                        )
                         User.permission_id = newPermission
                     else:
                         return ResponseJson({"status": 0, "msg": "所选择的权限组不存在"})
-                if disable and disable != User.disable:
-                    writeAudit(req.session.get("userID"), "Edit User Info(编辑用户): Edit Status(更新用户状态)",
-                               "User Manager(权限管理)",
-                               f"{not User.disable}-->{not disable}")
+                if disable is not None and disable != User.disable:
+                    writeAudit(
+                        req.session.get("userID"),
+                        "Edit User Info(编辑用户): Disable User(禁用用户)",
+                        "User Manager(用户管理)",
+                        f"{User.disable}-->{disable}"
+                    )
                     User.disable = disable
                 User.save()
                 return ResponseJson({"status": 1, "msg": "成功", "data": {
@@ -256,8 +287,11 @@ def setUserInfo(req):
                     "userName": User.userName,
                     "realName": User.realName,
                     "email": User.email,
-                    "permissionId": User.permission_id.id if isinstance(User.permission_id, Permission_groups) else User.permission_id,
-                    "permissionName": Permission_groups.objects.filter(id=User.permission_id.id if isinstance(User.permission_id, Permission_groups) else User.permission_id).first().name if User.permission_id else None,
+                    "permissionId": User.permission_id.id if isinstance(User.permission_id,
+                                                                        Permission_groups) else User.permission_id,
+                    "permissionName": Permission_groups.objects.filter(
+                        id=User.permission_id.id if isinstance(User.permission_id,
+                                                               Permission_groups) else User.permission_id).first().name if User.permission_id else None,
                     "disable": User.disable
                 }})
             else:
